fix(mobile): route Edge TTS through paired desktop instead of Metro dev proxy
Edge TTS on production web (app.dotagents.app) was 405-ing because the
dev-only /edge-tts proxy only exists when Metro is serving. Browsers
can't set the custom Origin/User-Agent headers that Microsoft's Edge
TTS WebSocket requires, so a direct WSS connection from the web build
isn't an option either.

This moves Edge TTS — and TTS in general for mobile — onto the paired
desktop's remote server.

Desktop:
- New apps/desktop/src/main/tts-service.ts with a provider-agnostic
  generateTTS(input, config) extracted from tipc.ts. Covers openai,
  groq, gemini, edge, kitten, supertonic.
- tipc.ts generateSpeech is now a thin wrapper.
- New POST /v1/tts/speak on remote-server.ts (Bearer-auth, 32KB text
  cap, returns raw audio bytes with the provider's Content-Type).

Mobile:
- Rename lib/edgeTts.ts -> lib/remoteTts.ts. Single path now: POST to
  ${baseUrl}/tts/speak, play via Audio on web or expo-audio on native.
- TTSSettings, SettingsScreen, ResponseHistoryPanel, ChatScreen all
  thread baseUrl + apiKey through and call speakRemoteTts.
- Edge voice picker is gated on pairing; when no desktop is paired,
  the picker hides Edge voices and playback falls back to native
  expo-speech.
- normalizeStoredConfig downgrades ttsProvider 'edge' -> 'native' when
  baseUrl/apiKey are missing, so a stale config from a previous pair
  doesn't try to speak through a desktop that isn't there.
- Metro /edge-tts dev proxy + script deleted (no longer called).
