+++ apps/desktop/src/main/tipc.ts
@@ -71,7 +71,7 @@ import {
   PanelPosition,
 } from "./panel-position"
 import { state, agentProcessManager, suppressPanelAutoShow, isPanelAutoShowSuppressed, toolApprovalManager, agentSessionStateManager } from "./state"
-import { generateEdgeTTS, type TTSGenerationResult } from "./edge-tts"
+import { generateTTS } from "./tts-service"
 
 
 import { startRemoteServer, stopRemoteServer, restartRemoteServer, printQRCodeToTerminal, getRemoteServerStatus } from "./remote-server"
@@ -566,8 +566,7 @@ import { summarizationService } from "./summarization-service"
 import { updateTrayIcon } from "./tray"
 import { isAccessibilityGranted } from "./utils"
 import { writeText, writeTextWithFocusRestore } from "./keyboard"
-import { preprocessTextForTTS, validateTTSText } from "@dotagents/shared"
-import { preprocessTextForTTSWithLLM } from "./tts-llm-preprocessing"
+
 
 
 const t = tipc.create()
@@ -3278,95 +3277,8 @@ export const router = {
       speed?: number
     }>()
     .action(async ({ input }) => {
-
-
-
-
-
-
-
-
-
-      const config = configStore.get()
-
-
-
-      if (!config.ttsEnabled) {
-        throw new Error("Text-to-Speech is not enabled")
-      }
-
-      const providerId = input.providerId || config.ttsProviderId || "openai"
-
-      // Preprocess text for TTS
-      let processedText = input.text
-
-      if (config.ttsPreprocessingEnabled !== false) {
-        // Use LLM-based preprocessing if enabled, otherwise fall back to regex
-        if (config.ttsUseLLMPreprocessing) {
-          processedText = await preprocessTextForTTSWithLLM(input.text, config.ttsLLMPreprocessingProviderId)
-        } else {
-          // Use regex-based preprocessing
-          const preprocessingOptions = {
-            removeCodeBlocks: config.ttsRemoveCodeBlocks ?? true,
-            removeUrls: config.ttsRemoveUrls ?? true,
-            convertMarkdown: config.ttsConvertMarkdown ?? true,
-          }
-          processedText = preprocessTextForTTS(input.text, preprocessingOptions)
-        }
-      }
-
-      // Validate processed text
-      const validation = validateTTSText(processedText)
-      if (!validation.isValid) {
-        throw new Error(`TTS validation failed: ${validation.issues.join(", ")}`)
-      }
-
       try {
-        let ttsResult: TTSGenerationResult
-
-
-
-        if (providerId === "openai") {
-          ttsResult = await generateOpenAITTS(processedText, input, config)
-        } else if (providerId === "groq") {
-          ttsResult = await generateGroqTTS(processedText, input, config)
-        } else if (providerId === "gemini") {
-          ttsResult = await generateGeminiTTS(processedText, input, config)
-        } else if (providerId === "edge") {
-          ttsResult = await generateEdgeTTS(processedText, input, config)
-        } else if (providerId === "kitten") {
-          const { synthesize } = await import('./kitten-tts')
-          const voiceId = config.kittenVoiceId ?? 0 // Default to Voice 2 - Male
-          const result = await synthesize(processedText, voiceId, input.speed)
-          const wavBuffer = float32ToWav(result.samples, result.sampleRate)
-          ttsResult = {
-            audio: new Uint8Array(wavBuffer).buffer,
-            mimeType: "audio/wav",
-          }
-        } else if (providerId === "supertonic") {
-          const { synthesize } = await import('./supertonic-tts')
-          const voice = config.supertonicVoice ?? "M1"
-          const lang = config.supertonicLanguage ?? "en"
-          const speed = input.speed ?? config.supertonicSpeed ?? 1.05
-          const steps = config.supertonicSteps ?? 5
-          const result = await synthesize(processedText, voice, lang, speed, steps)
-          const wavBuffer = float32ToWav(result.samples, result.sampleRate)
-          ttsResult = {
-            audio: new Uint8Array(wavBuffer).buffer,
-            mimeType: "audio/wav",
-          }
-        } else {
-          throw new Error(`Unsupported TTS provider: ${providerId}`)
-        }
-
-
-
-        return {
-          audio: ttsResult.audio,
-          mimeType: ttsResult.mimeType,
-          processedText,
-          provider: providerId,
-        }
+        return await generateTTS(input, configStore.get())
       } catch (error) {
         diagnosticsService.logError("tts", "TTS generation failed", error)
         throw error
@@ -5178,217 +5090,6 @@ export const router = {
     loopService.stopAllLoops()
     return { success: true }
   }),
-}
-
-// TTS Provider Implementation Functions
-
-function getOpenAITTSMimeType(responseFormat: "mp3" | "opus" | "aac" | "flac" | "wav" | "pcm"): string {
-  switch (responseFormat) {
-    case "mp3":
-      return "audio/mpeg"
-    case "opus":
-      return "audio/opus"
-    case "aac":
-      return "audio/aac"
-    case "flac":
-      return "audio/flac"
-    case "pcm":
-      return "audio/L16"
-    case "wav":
-    default:
-      return "audio/wav"
-  }
-}
-
-async function generateOpenAITTS(
-  text: string,
-  input: { voice?: string; model?: string; speed?: number },
-  config: Config
-): Promise<TTSGenerationResult> {
-  const model = input.model || config.openaiTtsModel || "gpt-4o-mini-tts"
-  const voice = input.voice || config.openaiTtsVoice || "alloy"
-  const speed = input.speed || config.openaiTtsSpeed || 1.0
-  const responseFormat = config.openaiTtsResponseFormat || "mp3"
-
-  const baseUrl = config.openaiBaseUrl || "https://api.openai.com/v1"
-  const apiKey = config.openaiApiKey
-
-
-
-  if (!apiKey) {
-    throw new Error("OpenAI API key is required for TTS")
-  }
-
-  const requestBody = {
-    model,
-    input: text,
-    voice,
-    speed,
-    response_format: responseFormat,
-  }
-
-
-
-  const response = await fetch(`${baseUrl}/audio/speech`, {
-    method: "POST",
-    headers: {
-      "Authorization": `Bearer ${apiKey}`,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(requestBody),
-  })
-
-
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(`OpenAI TTS API error: ${response.statusText} - ${errorText}`)
-  }
-
-  const audioBuffer = await response.arrayBuffer()
-
-  return {
-    audio: audioBuffer,
-    mimeType: getOpenAITTSMimeType(responseFormat),
-  }
-}
-
-async function generateGroqTTS(
-  text: string,
-  input: { voice?: string; model?: string },
-  config: Config
-): Promise<TTSGenerationResult> {
-  const model = input.model || config.groqTtsModel || "canopylabs/orpheus-v1-english"
-  // Choose default voice based on model - Arabic model should use Arabic voice
-  const defaultVoice = model === "canopylabs/orpheus-arabic-saudi" ? "fahad" : "troy"
-  const voice = input.voice || config.groqTtsVoice || defaultVoice
-
-  const baseUrl = config.groqBaseUrl || "https://api.groq.com/openai/v1"
-  const apiKey = config.groqApiKey
-
-
-
-  if (!apiKey) {
-    throw new Error("Groq API key is required for TTS")
-  }
-
-  const requestBody = {
-    model,
-    input: text,
-    voice,
-    response_format: "wav",
-  }
-
-
-
-  const response = await fetch(`${baseUrl}/audio/speech`, {
-    method: "POST",
-    headers: {
-      "Authorization": `Bearer ${apiKey}`,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(requestBody),
-  })
-
-
-
-  if (!response.ok) {
-    const errorText = await response.text()
-
-    // Check for specific error cases and provide helpful messages
-    if (errorText.includes("requires terms acceptance")) {
-      // The model parameter determines which terms page to show
-      const modelParam = model === "canopylabs/orpheus-arabic-saudi"
-        ? "canopylabs%2Forpheus-arabic-saudi"
-        : "canopylabs%2Forpheus-v1-english"
-      throw new Error(`Groq TTS model requires terms acceptance. Please visit https://console.groq.com/playground?model=${modelParam} and accept the terms when prompted, then try again.`)
-    }
-
-    throw new Error(`Groq TTS API error: ${response.statusText} - ${errorText}`)
-  }
-
-  const audioBuffer = await response.arrayBuffer()
-
-  return {
-    audio: audioBuffer,
-    mimeType: "audio/wav",
-  }
-}
-
-async function generateGeminiTTS(
-  text: string,
-  input: { voice?: string; model?: string },
-  config: Config
-): Promise<TTSGenerationResult> {
-  const model = input.model || config.geminiTtsModel || "gemini-2.5-flash-preview-tts"
-  const voice = input.voice || config.geminiTtsVoice || "Kore"
-
-  const baseUrl = config.geminiBaseUrl || "https://generativelanguage.googleapis.com"
-  const apiKey = config.geminiApiKey
-
-  if (!apiKey) {
-    throw new Error("Gemini API key is required for TTS")
-  }
-
-  const requestBody = {
-    contents: [{
-      parts: [{ text }]
-    }],
-    generationConfig: {
-      responseModalities: ["AUDIO"],
-      speechConfig: {
-        voiceConfig: {
-          prebuiltVoiceConfig: {
-            voiceName: voice
-          }
-        }
-      }
-    }
-  }
-
-  const url = `${baseUrl}/v1beta/models/${model}:generateContent?key=${apiKey}`
-
-
-
-  const response = await fetch(url, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(requestBody),
-  })
-
-
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(`Gemini TTS API error: ${response.statusText} - ${errorText}`)
-  }
-
-  const result = await response.json()
-
-
-
-  const inlineAudioData = result.candidates?.[0]?.content?.parts?.[0]?.inlineData
-  const audioData = inlineAudioData?.data
-
-  if (!audioData) {
-    throw new Error("No audio data received from Gemini TTS API")
-  }
-
-  // Convert base64 to ArrayBuffer
-  const binaryString = atob(audioData)
-  const bytes = new Uint8Array(binaryString.length)
-  for (let i = 0; i < binaryString.length; i++) {
-    bytes[i] = binaryString.charCodeAt(i)
-  }
-
-
-
-  return {
-    audio: bytes.buffer,
-    mimeType: inlineAudioData?.mimeType || "audio/L16",
-  }
 }
 
 export type Router = typeof router
